scripts/requirements.txt
pymbr
openocd
pyparted
humanfriendly
